+++ HISTORY.md
@@ -1,10 +1,26 @@
 # History
 
+## 4.2
+
+* Implement `includes`.
+
+## 4.1
+
+* Implement `equal`.
+
+## 4.0
+
+* Breaking changes:
+  * Rename `unemptyString` => `nonEmptyString`.
+  * Support derived error objects in `error`.
+* Fix HTMLElement instance predicate bug in Safari.
+
 ## 3.3
 
 * Implement `greaterOrEqual`.
 * Implement `lessOrEqual`.
 * Implement `inRange`.
+* Fix default error message for `function`. Thanks to [Paul Jolly][myitcv].
 
 ## 3.2
 
@@ -106,4 +122,5 @@
 [rodrigo]: https://github.com/roro89
 [ialex]: https://github.com/ialex
 [clintonskakun]: https://github.com/clintonskakun
+[myitcv]: https://github.com/myitcv